# api/openapi-main.yaml
openapi: 3.0.3
info:
  title: Agent Factory API
  version: 0.1.0
  license:
    name: MIT
  description: |
    Current Agent Factory runtime API contract. This file captures existing
    handler behavior for schema standardization and intentionally does not
    redesign route names, pagination, or response fields.
servers:
  - url: http://localhost:7437
    description: Local Agent Factory API server
security: []
...
paths:
...
  /models:
    get:
      tags:
        - Models
      operationId: listModels
      summary: List managed runtimes
      description: Lists named managed runtimes exposed by the currently loaded runtime configuration together with readiness, lifecycle state, locality, supported operations, and resource summary data using the managed-runtime contract.
      responses:
        '200':
          description: Managed runtimes for the current runtime.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListModelsResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /models/{model_name}:
    get:
      tags:
        - Models
      operationId: getModel
      summary: Inspect one managed runtime
      description: Returns one managed runtime's readiness, lifecycle state, supported operations, resource metadata, worker capabilities, and diagnostics for the currently loaded runtime configuration.
      parameters:
        - name: model_name
          in: path
          required: true
          description: Stable managed runtime identity such as `OMNIVOICE_Q4_K_M`.
          schema:
            type: string
      responses:
        '200':
          description: Managed runtime detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /models/{model_name}/invocations:
    post:
      tags:
        - Models
      operationId: invokeModel
      summary: Invoke one discovered model directly
      description: Invokes one discovered model through its declared provider-agnostic operation contract using canonical `WorkContent` input and optional slot-binding overrides. Non-streaming clients receive JSON metadata, while audio-producing operations can return a streamed audio body when requested.
      parameters:
        - name: model_name
          in: path
          required: true
          description: Concrete public model identifier such as `OMNIVOICE_Q4_K_M`.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelInvocationRequest'
      responses:
        '200':
          description: Successful model invocation result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInvocationResponse'
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Streamed audio body for audio-producing invocations when `options.responseMode` requests `AUDIO_STREAM`.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /models/{model_name}/pull:
    post:
      tags:
        - Models
      operationId: pullModel
      summary: Pull or install one managed runtime
      description: Pulls or installs required managed runtime assets into the managed cache using one source-agnostic customer action. Cloud-backed runtimes and unsupported local targets return actionable errors instead of silently succeeding.
      parameters:
        - name: model_name
          in: path
          required: true
          description: Stable managed runtime identity such as `OMNIVOICE_Q4_K_M`.
          schema:
            type: string
      responses:
        '200':
          description: Successful managed runtime pull or install, or confirmation that the managed cache already contained the required revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPullResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /provider-sessions/detail:
    get:
      tags:
        - Provider Sessions
      operationId: getProviderSessionDetails
      summary: Get provider session details
      description: Returns parsed provider-session details using a provider-neutral response schema. The browser supplies provider, kind, and identifier query parameters; the server resolves the matching session file under the configured provider sessions root and never accepts a raw filesystem path. Only Codex (`codex`) sessions are currently loadable for this endpoint.
      parameters:
        - name: provider
          in: query
          required: true
          description: Provider that emitted the session identifier. Only codex sessions are currently loadable.
          schema:
            $ref: '#/components/schemas/LoadableProviderSessionProvider'
        - name: kind
          in: query
          required: true
          description: Provider-session identifier kind. Only session_id is currently loadable.
          schema:
            $ref: '#/components/schemas/LoadableProviderSessionKind'
        - name: id
          in: query
          required: true
          description: Provider-session identifier to resolve. This is an identifier, not a filesystem path.
          schema:
            type: string
      responses:
        '200':
          description: Parsed provider-session details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderSessionDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
...
  responses:
    BadRequest:
      $ref: './components/responses/BadRequest.yaml'
...
    NotFound:
      $ref: './components/responses/NotFound.yaml'
...
    InternalError:
      $ref: './components/responses/InternalError.yaml'
...
  schemas:
...
    ListModelsResponse:
      $ref: './components/schemas/api/ListModelsResponse.yaml'
...
    ModelDetail:
      $ref: './components/schemas/api/ModelDetail.yaml'
    ModelInvocationRequest:
      $ref: './components/schemas/api/ModelInvocationRequest.yaml'
...
    ModelInvocationResponse:
      $ref: './components/schemas/api/ModelInvocationResponse.yaml'
...
    ModelPullResponse:
      $ref: './components/schemas/api/ModelPullResponse.yaml'
...
    ProviderSessionDetailResponse:
      $ref: './components/schemas/api/ProviderSessionDetailResponse.yaml'
...
    LoadableProviderSessionProvider:
      $ref: './components/schemas/api/LoadableProviderSessionProvider.yaml'
    LoadableProviderSessionKind:
      $ref: './components/schemas/api/LoadableProviderSessionKind.yaml'
...